app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
openai
google.auth.transport.requests
googleapiclient.discovery
google_auth_oauthlib.flow
bs4
python-dateutil